pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "aiassistant"
version = "0.1.0"
description = ""
authors = [
    {name = "Щиголь Алексей",email = "hosglial@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-docx (>=1.2.0,<2.0.0)",
    "ollama (>=0.6.1,<0.7.0)",
    "python-telegram-bot (>=22.5,<23.0)",
    "qdrant-client (>=1.7.0,<2.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "pydantic (>=2.12.4,<3.0.0)",
    "pydantic-settings (>=2.0.0,<3.0.0)",
    "asyncpg (>=0.29.0,<1.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
